requirements.txt
aiofiles==23.1.0
aiohttp==3.8.1
khl.py==0.3.7
